Feat: Resources

+++ JeekparkEngine_Window/jkTitleScene.cpp
@@ -21,7 +21,7 @@ namespace jk
         Scene::Update();
     }
     void TitleScene::LateUpdate()
-    {
+    { 
         Scene::LateUpdate();
         if (Input::GetKeyDown(eKeyCode::N))
         {
@@ -32,6 +32,6 @@ namespace jk
     {
         Scene::Render(hdc);
         wchar_t str[50] = L"Title Scene";
-        TextOut(hdc, 0, 0, str, wcsnlen_s(str, 50));
+        TextOut(hdc, 0, 0, str, (int)wcsnlen_s(str, 50));
     }
 }
